BZ #1493046 - Slow start of rails environment when starting passenger or foreman-tasks
Fixes #21295 - load the routes before we load the controllers

After upgrade from Rails 4.1 to Rails 4.2 there was significant
drop on rails startup in production. I have tracked it down to
several changes in Rails, but it might also be affected by the
fact the we use the url_helpers on more places than older Foreman
versions.

The reason of the slowness is the `define_method` calls on a module for
every named route we have. The `define_method` is more expensive
the more classes include the module that gets the new method.

The solution of this problem is to load the routes before we
load the controllers (including the eager loading) so that the modules
get defined new methods before the controllers include it in their code.

The original after_finalize loading is still here, but should not get
run again, as it callse `execute_if_updated` and nothing new usually
gets updated there.

To test, one can try in devel setup:

    time echo exit | bundle exec rails c production

or in production:

    service httpd restart; time curl -k https://localhost

In devel setup, I got from 30s to 10s, in production with Katello
and other plugins (it gets more and more visible with more plugins
around), I got from 1m 15s to 15s.

Please not `foreman-rake console` is not affected, as the rake
environment turns off the eager loading.

(cherry picked from commit c2e58b1b275e0107de4089e15adc8b6203e696b1)