BZ #1513932 - OpenStack v3 keystone authentication
Fixes #12054 - Openstack v3 support

Fixes #21682 - Add Domain to Compute Resource API (OpenStack)

The Domain field was added to the Compute Resources table in #12054 as
part of the Openstack v3 support. This field needs to appear in the
APIdoc to be able to use it and in the RABL to read it.

Fixes #21675 - OpenStack create host tab fails to load

Currently the tab is not loading as there are a few methods that are not
being overridden properly by the FogExtension.

After fixing that and adding the metadata key the tab loads