BZ #1204667 - Allow unicode in mail addresses
The version of the mail gem we ship with Satellite incorrectly displays
the address as encoded leading to the validator failing. This will be
fixed upstream by bumping the version of the mail gem, this is a
downstream-only fix to correctly decode the address using the current
version of the gem.